backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Astronomical Calculations
skyfield==1.48
numpy==1.26.3
pandas==2.2.0
pyswisseph==2.10.3.2  # Swiss Ephemeris for astrology calculations
pytz==2024.1  # Timezone support for astrology

# AI & LLM
google-generativeai==0.8.3  # Gemini AI for astrology commentary

# Data Processing
astropy==6.0.0
astroquery==0.4.7

# Database
aiosqlite==0.19.0

# Visualization
matplotlib==3.8.2
Pillow==10.2.0

# Geospatial Data (for NASA VNL light pollution)
rasterio==1.3.11
pyproj==3.6.1

# PDF Generation
reportlab==4.0.9
weasyprint==60.2

# HTTP Clients
httpx==0.26.0
requests==2.31.0

# Utilities
python-dotenv==1.0.0
cachetools==5.3.2

